requirements.txt
kafka-python==1.4.7
oauthlib==3.1.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
configparser==5.0.0
